[#1380] Extract other methods out of set_global_secret
Extracts `load_fortunes`, `check_global_banner`,
  and `load_plans` methods from `set_global_secret`
  into separate initializer files.
